# Introduce Range.updatePoints