#### Is this adding or improving a _feature_ or fixing a _bug_?

feature

#### What's the new behavior?

Introduce `Range.updatePoints(Point => Point)` to make the code simpler.

#### How does this change work?

We used to have `range.setPoints([anchor.doSomething, focus.doSomething])` that same point operation duplicated in setPoint.  This function mimics Immutable's update to reduce duplicated code.

#### Have you checked that...?

<!-- 
Please run through this checklist for your pull request: 
-->

* [x] The new code matches the existing patterns and styles.
* [x] The tests pass with `yarn test`.
* [x] The linter passes with `yarn lint`. (Fix errors with `yarn prettier`.)
* [x] The relevant examples still work. (Run examples with `yarn watch`.)

#### Does this fix any issues or need any specific reviewers?

Fixes: #
Reviewers: @