@zhujinxuan 2018-08-14
Hi, @ianstormtaylor .  Just by using withMutations.  The mapRanges is speeded up significantly

```
  benchmarks
    changes
      ✔ add-mark: 84.49 -> 90.89 ops/sec (7.57% faster)
      ✔ delete-backward: 110.78 -> 112.90 ops/sec (1.91% faster)
      ✔ delete-forward: 123.53 -> 122.62 ops/sec (1.00% slower)
      ✔ insert-node-by-key: 182.38 -> 182.00 ops/sec (0.00% slower)
      ✔ insert-text-by-key: 217.83 -> 201.21 ops/sec (8.00% slower)
      ✔ insert-text: 126.27 -> 129.68 ops/sec (2.70% faster)
      ✔ normalize: 147.59 -> 147.37 ops/sec (0.00% slower)
      ✔ remove-node-by-key: 136.30 -> 132.21 ops/sec (3.00% slower)
      ✔ set-blocks: 172.70 -> 184.56 ops/sec (6.86% faster)
      ✔ split-block: 65.79 -> 68.07 ops/sec (3.47% faster)
    models
      ✔ from-json-big: 4.67 -> 4.73 ops/sec (1.34% faster)
      ✔ from-json: 262.27 -> 295.22 ops/sec (12.56% faster)
      ✔ get-active-marks-at-range: 2842.08 -> 2794.84 ops/sec (2.00% slower)
      ✔ get-ancestors: 2810.04 -> 2612.03 ops/sec (8.00% slower)
      ★ get-blocks-at-range:
            user: 569.81 -> 659.08 ops/sec (15.67% faster)
            real: 901.12 -> 1296.50 ops/sec (43.88% faster)
      ★ get-blocks:
            user: 7054.60 -> 9296.21 ops/sec (31.78% faster)
            real: 8980.15 -> 15076.53 ops/sec (67.89% faster)
      ✔ get-common-ancestor: 5774.97 -> 5945.55 ops/sec (2.95% faster)
      ★ get-descendant:
            user: 3995.10 -> 4970.57 ops/sec (24.42% faster)
            real: 3439.36 -> 6579.39 ops/sec (91.30% faster)
      ★ get-furthest-ancestor:
            user: 5615.88 -> 7008.76 ops/sec (24.80% faster)
            real: 4578.65 -> 8256.18 ops/sec (80.32% faster)
      ★ get-inlines-at-range:
            user: 981.87 -> 1034.35 ops/sec (5.34% faster)
            real: 1011.01 -> 1770.46 ops/sec (75.12% faster)
      ✔ get-inlines: 14292.23 -> 13277.57 ops/sec (8.00% slower)
      ✔ get-leaves: 1192342.57 -> 1493400.60 ops/sec (25.25% faster)
      ✔ get-marks-at-range: 675.04 -> 780.53 ops/sec (15.63% faster)
      ✔ get-marks: 1714.13 -> 2171.97 ops/sec (26.71% faster)
      ✔ get-parent: 5024.91 -> 4662.11 ops/sec (8.00% slower)
      ✔ get-path: 6542.93 -> 6506.75 ops/sec (1.00% slower)
      ✔ get-texts-at-range: 3131.93 -> 3024.24 ops/sec (4.00% slower)
      ★ get-texts:
            user: 12969.97 -> 17060.46 ops/sec (31.54% faster)
            real: 16204.02 -> 23398.23 ops/sec (44.40% faster)
      ✔ has-node-multiple: 3538.61 -> 3427.70 ops/sec (3.00% slower)
      ★ has-node:
            user: 4320.14 -> 5060.81 ops/sec (17.14% faster)
            real: 4580.03 -> 6596.36 ops/sec (44.02% faster)
      ★ to-json:
            user: 2103.15 -> 2645.79 ops/sec (25.80% faster)
            real: 2886.63 -> 5277.50 ops/sec (82.83% faster)
    html-serializer
      ✔ deserialize: 23.24 -> 24.70 ops/sec (6.29% faster)
      ★ serialize:
            user: 283.47 -> 480.08 ops/sec (69.36% faster)
            real: 390.46 -> 866.91 ops/sec (122.02% faster 🙌 )
    plain-serializer
      ✔ deserialize: 1182.20 -> 1360.67 ops/sec (15.10% faster)
      ★ serialize:
            user: 3730.53 -> 4999.66 ops/sec (34.02% faster)
            real: 6764.43 -> 8599.31 ops/sec (27.13% faster)
    rendering
      ✔ decoration: 69.95 -> 77.21 ops/sec (10.37% faster)
      ✔ normal: 77.68 -> 92.52 ops/sec (19.10% faster)

✨  Done in 64.73s.
```