@zhujinxuan 2018-08-14
@ianstormtaylor Hi, I am not sure about the test `insert-node-by-key/duplicate`.  It seems we shall never insert two nodes with the same key?