@zhujinxuan 2018-08-14
Await #2081 to see how to resolve the duplicate key issue in reset path.