@ianstormtaylor 2018-08-16
@zhujinxuan very cool! I love how it mimics the existing Immutable.js `update` method. Great call!